@@ -16,7 +16,7 @@ Gem::Specification.new do |s|
 
   s.files = Dir["{app,config,db,lib}/**/*", "MIT-LICENSE", "Rakefile", "README.md"]
 
-  s.add_dependency "rails", ">= 5.2"
+  s.add_dependency "rails", "~> 6.0"
   s.add_dependency 'cancancan'
   s.add_dependency 'paperclip', "~> 6.0"
   s.add_dependency "paperclip-av-transcoder"
