@@ -55,7 +55,7 @@ module Chemistry
     end
   
     def update
-      if @page_collection.update_attributes(page_collection_params)
+      if @page_collection.update(page_collection_params)
         redirect_to page_collection_home_url(@page_collection.slug)
       else
         render action: :edit
@@ -67,7 +67,7 @@ module Chemistry
     end
   
     def create
-      if @page_collection.update_attributes(page_collection_params)
+      if @page_collection.update(page_collection_params)
         redirect_to page_collection_home_url(@page_collection)
       else
         render action: :new
