@@ -28,7 +28,11 @@ class Chemistry::PageSerializer
              :published_at
 
   attribute :url do |object|
-    Chemistry::Engine.routes.url_helpers.published_page_url(object.path, host: Chemistry.config.site_host)
+    if object.home?
+      Chemistry::Engine.routes.url_helpers.home_page_url(host: Chemistry.config.site_host)
+    else
+      Chemistry::Engine.routes.url_helpers.published_page_url(object.path, host: Chemistry.config.site_host)
+    end
   end
 
 end
